+++ src/front/js/layout.jsx
@@ -4,14 +4,12 @@ import ScrollToTop from "./component/scrollToTop.jsx";
 
 import injectContext from "./store/appContext";
 
-import { Home } from "./views/home.jsx";
-import { Demo } from "./views/demo.jsx";
-import LoginPage from "./views/LoginPage/LoginPage.jsx";
 
-import { Navbar } from "./component/navbar.jsx";
-import { Footer } from "./component/footer.jsx";
+import LoginPage from "./views/Login/LoginComponents/LoginPage.jsx";
+
+
 import TodoList from "./views/DragDrop/todoList.jsx";
-import Register from "./views/LoginPage/AccountApp.jsx";
+import Register from "./views/Login/RegisterComponents/AccountApp.jsx";
 //create your first component
 const Layout = () => {
   //the basename is used when your project is published in a subdirectory and not in the root of the domain
@@ -30,13 +28,15 @@ const Layout = () => {
             <Route exact path="/AccountApp">
               <Register />
             </Route>
+            <Route exact path="/LoginApp">
+              <LoginPage />
+              </Route>
             <Route exact path="/todoList">
               <TodoList />
             </Route>
             <Route exact path="/">
               <LoginPage />
             </Route>
-
             {/* add more routes when files merge */}
           </Switch>
         </ScrollToTop>
