+++ src/front/js/views/Login/RegisterComponents/RegisterApp.jsx
@@ -1,0 +1,9 @@
+import React from 'react'
+
+const RegisterApp = () => {
+  return (
+    <div>RegisterApp</div>
+  )
+}
+
+export default RegisterApp
