+++ src/front/js/views/Login/css/register.css
@@ -1,0 +1,60 @@
+.register-container {
+  height: 600px;
+  width: 500px;
+  border-radius: 30px;
+  background-color: rgba(255, 253, 253, 0.952);
+  margin: auto;
+  position: absolute;
+  top: 0;
+  bottom: 0;
+  left: 0;
+  right: 0;
+  font-family: "DM Sans", sans-serif;
+  font-weight: 600;
+  text-align: center;
+}
+.buddy-logo{
+  color: rgba(255, 255, 255, 0.5);
+  width: 200px;
+  border-radius: 5px;
+}
+.content {
+  text-align: center;
+  padding-top: 20px;
+}
+h2 {
+  margin-top: 40px;
+}
+.form-group {
+  margin-top: 15px;
+}
+.reg-user {
+  margin-bottom: 20px;
+  width: 380px;
+  height: 45px;
+  border-radius: 10px;
+  font-weight: 600;
+  border: none;
+}
+
+
+.btn {
+  font-size: 20px;
+  letter-spacing: 0.1em;
+  width: 380px;
+  height: 45px;
+  position: relative;
+  left: 0px;
+  top: 30px;
+  color: #3d3a3a;
+  background-color: #dbdbdb;
+  border-radius: 10px;
+  border: none;
+  /* border-color: #ebbdd8; */
+}
+
+.footer:hover {
+  transform: scale(1.04);
+  transition: 0.3s ease-in-out;
+  cursor: pointer;
+}
