--- src/front/js/views/LoginPage/register.css
@@ -1,15 +1,0 @@
-.register-container {
-  height: 550px;
-  width: 500px;
-  border-radius: 30px;
-  background-color: rgba(255, 253, 253, 0.952);
-  margin: auto;
-  position: absolute;
-  top: 0;
-  bottom: 0;
-  left: 0;
-  right: 0;
-  font-family: "DM Sans", sans-serif;
-  font-weight: 600;
-  text-align: center;
-}
